ml_pipeline/requirements.txt
imbalanced_learn==0.14.0
joblib==1.5.2
loguru==0.7.3
matplotlib==3.9.0
numpy==1.26.4
optuna==4.4.0
pandas==2.1.1
python_box==7.3.2
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
sdmetrics==0.23.0
sdv==1.25.0
seaborn==0.13.2
setuptools==80.9.0
tqdm==4.67.1
xgboost==3.1.1